Fixes building libraries
