+++ .circleci/checksum.txt
@@ -1,13 +1,12 @@
-95482048463cdc99c77a69951654de7b28fd6880  ../Projects/Argon2/Argon2.csproj
-ebafae4e8160057cbf408cb13114479ca26e1b58  ../Projects/Scripts.Tests/Scripts.Tests.csproj
-6c6a48220658bdfa7b2e1e173077fd139ed4af2c  ../Projects/Scripts/Scripts.csproj
-6213c10f697b45729bab15f7870c42dcf49b026b  ../Projects/Server.Tests/Server.Tests.csproj
-4b3626e04f76dff1ab445e207658a3b386fce4e5  ../Projects/Server/Server.csproj
-a663d309b6b86c7632ea5cd175468dd5180808c1  ../Projects/Argon2/Argon2.cs
+6b0276b603e867108643104978963ddd2e599e58  ../Projects/Argon2/Argon2.csproj
+214715e88298b66715dee7c06080fc083f2c9e26  ../Projects/Argon2/Argon2.cs
 1e89625abe364f895346d0c5753f3ca6653cf7df  ../Projects/Argon2/Argon2Context.cs
 abbb34d85f55ca1d78a400b4d46afbd6965f642b  ../Projects/Argon2/Argon2Error.cs
 00ef35a6809d4959bdda1de01f45e39ff388270f  ../Projects/Argon2/Argon2Exception.cs
-53164d1621e24adefc6917379d36d7273551f457  ../Projects/Argon2/Argon2PasswordHasher.cs
+58979ca359777ae45a6233264f0741ba23203959  ../Projects/Argon2/Argon2PasswordHasher.cs
 a137d2fe0478d065d478366168f80ab9eaaa9c3a  ../Projects/Argon2/Argon2Type.cs
 c09f7a2f9690d56ca838aafb28640c09950c5a8c  ../Projects/Argon2/HashMetadata.cs
-b5da07548eb41ec9bd35bc2def006f23698b357e  ../Projects/Argon2/Utility.cs
+c9fb8b37bd035c6597047d4279f9b6e6384db944  ../Projects/Argon2/Utility.cs
+c9fb8b37bd035c6597047d4279f9b6e6384db944  ../Projects/ZLib/Utility.cs
+a4c3595dada3e596ab3131886616d5c27107ec7f  ../Projects/ZLib/ZLib.cs
+d129d82f8607e97090088a6e1956706d886e8294  ../Projects/ZLib/ZLib.csproj
