+++ .circleci/config.yml
@@ -21,15 +21,15 @@ jobs:
           keys:
             - dotnet-packages-v1-{{ arch }}-{{ checksum ".circleci/checksum.txt" }}
       - run:
-          name: Build Argon2
-          command: dotnet build -c Release -r linux-x64 Projects/Argon2/Argon2.csproj
-      - run:
-          name: Install ModernUO dependencies
-          command: dotnet restore
+          name: Build native libraries
+          command: /mnt/ramdisk/Tools/build-native-libraries.sh
       - save_cache:
           paths:
             - /mnt/ramdisk/packages
           key: dotnet-packages-v1-{{ arch }}-{{ checksum ".circleci/checksum.txt" }}
+      - run:
+          name: Install ModernUO dependencies
+          command: dotnet restore
       - run:
           name: Building ModernUO for Linux (Debug)
           command: dotnet build -c Debug -r linux-x64
@@ -42,15 +42,15 @@ jobs:
           keys:
             - dotnet-packages-v1-{{ arch }}-{{ checksum ".circleci/checksum.txt" }}
       - run:
-          name: Build Argon2
-          command: dotnet build -c Debug -r win-x64 Projects/Argon2/Argon2.csproj
+          name: Build native libraries
+          command: Tools\build-native-libraries.cmd
+      - save_cache:
+          paths:
+            - packages
+          key: dotnet-packages-v1-{{ arch }}-{{ checksum ".circleci/checksum.txt" }}
       - run:
           name: Install ModernUO dependencies
           command: dotnet restore
-      - save_cache:
-          paths:
-            - /mnt/ramdisk/packages
-          key: dotnet-packages-v1-{{ arch }}-{{ checksum ".circleci/checksum.txt" }}
       - run:
           name: Building ModernUO for Windows (Debug)
           command: dotnet build -c Debug -r win-x64
@@ -64,15 +64,15 @@ jobs:
           keys:
             - dotnet-packages-v1-{{ arch }}-{{ checksum ".circleci/checksum.txt" }}
       - run:
-          name: Build Argon2
-          command: dotnet build -c Debug -r linux-x64 Projects/Argon2/Argon2.csproj
-      - run:
-          name: Install ModernUO dependencies
-          command: dotnet restore
+          name: Build native libraries
+          command: /mnt/ramdisk/Tools/build-native-libraries.sh
       - save_cache:
           paths:
             - /mnt/ramdisk/packages
           key: dotnet-packages-v1-{{ arch }}-{{ checksum ".circleci/checksum.txt" }}
+      - run:
+          name: Install ModernUO dependencies
+          command: dotnet restore
       - run:
           name: Building with Analyzers
           command: dotnet build -c Analyze -r linux-x64
@@ -86,15 +86,15 @@ jobs:
           keys:
             - dotnet-packages-v1-{{ arch }}-{{ checksum ".circleci/checksum.txt" }}
       - run:
-          name: Build Argon2
-          command: dotnet build -c Debug -r linux-x64 Projects/Argon2/Argon2.csproj
-      - run:
-          name: Install ModernUO dependencies
-          command: dotnet restore
+          name: Build native libraries
+          command: /mnt/ramdisk/Tools/build-native-libraries.sh
       - save_cache:
           paths:
             - /mnt/ramdisk/packages
-          key: dotnet-packages-linux-{{ checksum ".circleci/checksum.txt" }}
+          key: dotnet-packages-v1-{{ arch }}-{{ checksum ".circleci/checksum.txt" }}
+      - run:
+          name: Install ModernUO dependencies
+          command: dotnet restore
       - run:
           name: Running unit tests
           command: dotnet test -r linux-x64
