+++ .circleci/genchecksum.sh
@@ -1,1 +1,1 @@
-sha1sum ../Projects/*/*.csproj ../Projects/Argon2/*.cs > "checksum.txt"
+sha1sum ../Projects/Argon2/*.csproj ../Projects/Argon2/*.cs ../Projects/ZLib/*.cs ../Projects/ZLib/*.csproj > "checksum.txt"
