+++ Projects/Scripts/Multis/HouseFoundation.cs
@@ -10,6 +10,7 @@ using Server.Items;
 using Server.Mobiles;
 using Server.Network;
 using Server.Spells;
+using Server.Compression;
 
 namespace Server.Multis
 {
@@ -2324,7 +2325,7 @@ namespace Server.Multis
         byte[] inflatedBuffer = m_PlaneBuffers[i];
 
         int deflatedLength = m_DeflatedBuffer.Length;
-        ZLibError ce = NetworkCompression.Pack(m_DeflatedBuffer, ref deflatedLength, inflatedBuffer, size,
+        ZLibError ce = ZLib.Pack(m_DeflatedBuffer, ref deflatedLength, inflatedBuffer, size,
           ZLibQuality.Default);
 
         if (ce != ZLibError.Okay)
@@ -2360,7 +2361,7 @@ namespace Server.Multis
         byte[] inflatedBuffer = m_StairBuffers[i];
 
         int deflatedLength = m_DeflatedBuffer.Length;
-        ZLibError ce = NetworkCompression.Pack(m_DeflatedBuffer, ref deflatedLength, inflatedBuffer, size,
+        ZLibError ce = ZLib.Pack(m_DeflatedBuffer, ref deflatedLength, inflatedBuffer, size,
           ZLibQuality.Default);
 
         if (ce != ZLibError.Okay)
