+++ Projects/Scripts/Scripts.csproj
@@ -66,7 +66,8 @@
     <PackageReference Include="Microsoft.AspNetCore.Connections.Abstractions" Version="3.1.3" />
     <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="3.1.3" />
     <PackageReference Include="Microsoft.Extensions.Hosting.Abstractions" Version="3.1.3" />
-    <PackageReference Include="Argon2.Bindings" Version="1.0.0" />
+    <PackageReference Include="Argon2.Bindings" Version="1.1.0" />
+    <PackageReference Include="Zlib.Bindings" Version="1.0.0" />
   </ItemGroup>
   <ItemGroup Condition="'$(Configuration)|$(Platform)'=='Analyze|AnyCPU'">
     <PackageReference Include="StyleCop.Analyzers" Version="1.1.118" PrivateAssets="all" />
