+++ Projects/Zlib/Zlib.csproj
@@ -1,9 +1,9 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
+<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <RuntimeIdentifiers>win-x64;linux-x64;osx-x64</RuntimeIdentifiers>
-    <PackageVersion>1.1.0</PackageVersion>
+    <PackageVersion>1.0.0</PackageVersion>
     <RootNamespace>Server</RootNamespace>
-    <AssemblyName>Argon2.Bindings</AssemblyName>
+    <AssemblyName>ZLib.Bindings</AssemblyName>
     <PlatformTarget>x64</PlatformTarget>
     <LangVersion>8.0</LangVersion>
     <TargetFramework>netcoreapp3.1</TargetFramework>
@@ -16,23 +16,23 @@
     <Platforms>x64</Platforms>
   </PropertyGroup>
   <ItemGroup>
-    <Content Include="runtimes\win-x64\native\argon2.dll">
+    <Content Include="runtimes\win-x64\native\zlib.dll">
       <Pack>true</Pack>
       <PackagePath>runtimes/win-x64/native</PackagePath>
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
-    <Content Include="runtimes\osx-x64\native\libargon2.dylib">
+    <Content Include="runtimes\osx-x64\native\libz.dylib">
       <Pack>true</Pack>
       <PackagePath>runtimes/osx-x64/native</PackagePath>
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
-    <Content Include="runtimes\linux-x64\native\libargon2.so">
+    <Content Include="runtimes\linux-x64\native\libz.so">
       <Pack>true</Pack>
       <PackagePath>runtimes/linux-x64/native</PackagePath>
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
   </ItemGroup>
   <Target Name="CopyPackage" AfterTargets="Pack">
-    <Exec Command="dotnet nuget push $(OutputPath)..\$(PackageId).$(PackageVersion).nupkg -s local" />
+    <Exec Command="dotnet nuget push $(OutputPath)..\..\$(PackageId).$(PackageVersion).nupkg -s local" />
   </Target>
 </Project>
