+++ Tools/build-native-libraries.cmd
@@ -1,0 +1,2 @@
+dotnet build -c Release Projects/Argon2/Argon2.csproj
+dotnet build -c Release Projects/ZLib/ZLib.csproj
